fix: consider erc20 decimals for the output token

+++ packages/react/src/modules/token/metadata/hooks/useContractDecimals.ts
@@ -13,9 +13,9 @@ export const useContractDecimals = (config: PredefinedReadContractConfig) => {
         name: 'decimals',
         outputs: [
           {
-            internalType: 'uint256',
+            internalType: 'uint8',
             name: '',
-            type: 'uint256',
+            type: 'uint8',
           },
         ],
         stateMutability: 'view',
